v0.0.303 structured output
