+++ lionagi/core/flow/monoflow.py
@@ -3,9 +3,10 @@ from typing import Any
 from lionagi.libs import ln_convert as convert
 from lionagi.libs import ln_func_call as func_call
 from lionagi.libs import ln_nested as nested
+from lionagi.libs.ln_parse import ParseUtil
 
 from lionagi.core.schema.base_node import Tool, TOOL_TYPE
-from lionagi.core.session.base.schema import Instruction, System
+from lionagi.core.messages.schema import Instruction
 
 
 class BaseMonoFlow:
@@ -25,6 +26,7 @@ class MonoChat(BaseMonoFlow):
 
     def __init__(self, branch) -> None:
         super().__init__(branch)
+
 
     def process_chatcompletion(self, payload, completion, sender):
         if "choices" in completion:
@@ -58,6 +60,7 @@ class MonoChat(BaseMonoFlow):
         tools=False,
         out: bool = True,
         invoke: bool = True,
+        output_fields=None,
         **kwargs,
     ) -> Any:
         """
@@ -79,7 +82,7 @@ class MonoChat(BaseMonoFlow):
         """
         if system:
             self.branch.change_first_system_message(system)
-        self.branch.add_message(instruction=instruction, context=context, sender=sender)
+        self.branch.add_message(instruction=instruction, context=context, sender=sender, output_fields=output_fields)
 
         if "tool_parsed" in kwargs:
             kwargs.pop("tool_parsed")
@@ -97,6 +100,7 @@ class MonoChat(BaseMonoFlow):
 
         async def _output():
             content_ = self.branch.last_message_content
+            
             if invoke:
                 try:
                     tool_uses = content_
@@ -121,13 +125,22 @@ class MonoChat(BaseMonoFlow):
                 except:
                     pass
             if out:
+                out_ = ''
                 if (
                     len(content_.items()) == 1
                     and len(nested.get_flattened_keys(content_)) == 1
                 ):
                     key = nested.get_flattened_keys(content_)[0]
-                    return content_[key]
-                return content_
+                    out_ = content_[key]
+                out_ = content_
+                
+                if output_fields:
+                    try:
+                        return ParseUtil.md_to_json(out_['response']) if 'response' in out_ else ParseUtil.md_to_json(out_)
+                    except:
+                        pass
+                
+                return out_
 
         return await _output()
