+++ lionagi/core/messages/schema.py
@@ -155,13 +155,30 @@ class Instruction(BaseMessage):
     """
 
     def __init__(
-        self, instruction: dict | list | str, context=None, sender: str | None = None
+        self, instruction: dict | list | str, context=None, sender: str | None = None, output_fields=None
     ):
         super().__init__(
             role="user", sender=sender or "user", content={"instruction": instruction}
         )
         if context:
             self.content.update({"context": context})
+
+        if output_fields:
+            format = """
+            Follow the following response format.
+            ```json
+            {                    
+            """
+            format += f"{output_fields}"
+            format += """
+            }
+            ```
+            """
+            self.content.update(
+                {
+                    "response_format": format
+                }
+            )
 
 
 class System(BaseMessage):
@@ -471,30 +488,3 @@ class Response(BaseMessage):
 #                    relation=relation, **kwargs)
 
 
-class MailCategory(str, Enum):
-    MESSAGES = "messages"
-    TOOL = "tools"
-    SERVICE = "provider"
-    MODEL = "model"
-
-
-class BaseMail:
-
-    def __init__(self, sender, recipient, category, package):
-        self.sender = sender
-        self.recipient = recipient
-        try:
-            if isinstance(category, str):
-                category = MailCategory(category)
-            if isinstance(category, MailCategory):
-                self.category = category
-            else:
-                raise ValueError(
-                    f"Invalid request title. Valid titles are" f" {list(MailCategory)}"
-                )
-        except Exception as e:
-            raise ValueError(
-                f"Invalid request title. Valid titles are "
-                f"{list(MailCategory)}, Error: {e}"
-            )
-        self.package = package
