+++ lionagi/tests/test_core/test_base_branch.py
@@ -1,9 +1,8 @@
-from lionagi.core.session.base.base_branch import BaseBranch
-from lionagi.core.session.base.util import MessageUtil, System
-from lionagi.core.schema import DataLogger
+from lionagi.core.branch.base_branch import BaseBranch
+from lionagi.core.branch.util import MessageUtil, System
 
 import unittest
-from unittest.mock import patch, MagicMock, create_autospec
+from unittest.mock import patch, MagicMock
 import pandas as pd
 from datetime import datetime
 import json
