+++ lionagi/tests/test_core/test_session_base_util.py
@@ -1,8 +1,7 @@
-from lionagi.core.session.base.util import MessageUtil
-from lionagi.core.session.base.schema import System, Instruction, Response
+from lionagi.core.branch.util import MessageUtil
+from lionagi.core.messages.schema import System, Instruction, Response
 
 import unittest
-from unittest.mock import patch
 import pandas as pd
 import json
 from datetime import datetime
