+++ lionagi/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.0.302"
+__version__ = "0.0.303"
